+++ guest/error_test.go
@@ -1,4 +1,4 @@
-package api
+package guest
 
 import (
 	"errors"
